accounts delete 시 confirm 기능 추가

+++ components/data.table.tsx
@@ -5,6 +5,7 @@ import * as React from "react"
 import { Trash } from "lucide-react"
 import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
+import { useConFirm } from "@/hooks/use-confirm"
 
 import {
   ColumnDef,
@@ -47,6 +48,11 @@ export function DataTable<TData, TValue>({
   disabled,
 }: DataTableProps<TData, TValue>) {
 
+  const [ConfirmDialog, confirm] = useConFirm(
+    " a u sure ?",
+    'u a about to perform a bulk delete'
+  )
+
   const [sorting, setSorting] = React.useState<SortingState>([])
   const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>(
     []
@@ -74,6 +80,7 @@ export function DataTable<TData, TValue>({
 
   return (
     <div>
+      <ConfirmDialog />
       <div className="flex items-center py-4">
         <Input
           placeholder={`Filter ${filterKey}...`}
@@ -89,13 +96,16 @@ export function DataTable<TData, TValue>({
             size="sm"
             variant="outline"
             className="ml-auto font-normal text-xs"
-            onClick={() => {
-              onDelete(table.getFilteredSelectedRowModel().rows)
-              table.resetRowSelection();
+            onClick={async () => {
+              const ok = await confirm();
+              if (ok) {
+                onDelete(table.getFilteredSelectedRowModel().rows)
+                table.resetRowSelection();
+              }
             }}
           >
             <Trash className="size-4 mr-2" />
-            Delete({table.getFilteredSelectedRowModel().rows.length})            
+            Delete({table.getFilteredSelectedRowModel().rows.length})
           </Button>
         )}
       </div>
